Add files via upload
These are the codes that resample and backtest the data. For resampling, it takes the price and market cap in the first week of each month (starting from Aug 2021) and puts them in the Monthly_Data csv. For basketsting, the code takes the price on the first week of each month and sums them together in the Weekly_sum_price file